finetuning integer division

+++ Booking.php
@@ -125,7 +125,7 @@ WEBSITE : HOMEPAGE OF A HOTEL'S WEBSITE
     $roomquery = $db->query("SELECT distinct r.`capacity` as capacity FROM `Room` r WHERE r.`type` = '$roomType'");
     $roomnumber_result = mysql_fetch_assoc($roomquery);
     $capacity = intval($roomnumber_result['capacity']);
-    $numberOfRoomsNeeded = intval($number)/$capacity;
+    $numberOfRoomsNeeded = (int)intval($number)/$capacity;
     if (intval($number)%$capacity > 0) {
       $numberOfRoomsNeeded = $numberOfRoomsNeeded + 1;
     }
